MakeBoxes[Infix[...],form]-> Infix
