+++ mathics/core/systemsymbols.py
@@ -170,6 +170,7 @@ SymbolRealDigits = Symbol("System`RealDigits")
 SymbolRepeated = Symbol("System`Repeated")
 SymbolRepeatedNull = Symbol("System`RepeatedNull")
 SymbolReturn = Symbol("System`Return")
+SymbolRight = Symbol("System`Right")
 SymbolRound = Symbol("System`Round")
 SymbolRow = Symbol("System`Row")
 SymbolRowBox = Symbol("System`RowBox")
